% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{change_name}
\alias{change_name}
\title{modify the distribution}
\usage{
change_name(replicate)
}
\arguments{
\item{replicate}{the distribution of PROTREC scores for the predicted, validated and original proteins}
}
\value{
output the modified variable
}
\description{
modify the distribution
}
