% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{complex_vector}
\alias{complex_vector}
\title{This is the complex_vector included in my package}
\description{
This is the complex_vector included in my package
}
\references{
\url{http://mips.helmholtz-muenchen.de/corum/}
}
