% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{fcs_cancer_mat}
\alias{fcs_cancer_mat}
\title{This is the fcs_cancer_mat dataset included in my package}
\description{
This is the fcs_cancer_mat dataset included in my package
}
