% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{fcs_prot_prob_assign}
\alias{fcs_prot_prob_assign}
\title{This assigns probabilities to individual proteins based on FCS}
\usage{
fcs_prot_prob_assign(cplx, p)
}
\arguments{
\item{cplx}{complex vector}

\item{p}{is the complex-based FCS probability, typically should be the 1 minus the result from certain sample of fcs}
}
\value{
The vector of FCS probabilities for the proteins of a sample
}
\description{
This assigns probabilities to individual proteins based on FCS
}
