% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{FCS_sig_complex_n}
\alias{FCS_sig_complex_n}
\title{This is the FCS_sig_complex_n dataset included in my package}
\description{
This is the FCS_sig_complex_n dataset included in my package
}
