% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_average_values_c}
\alias{get_average_values_c}
\title{This performs fcs-filtering of complexes before applying the PROTREC method using the RCC dataset}
\usage{
get_average_values_c(FCS_sig_complex_c)
}
\arguments{
\item{FCS_sig_complex_c}{A vector containing proteins associated with significant complexes in FCS}
}
\value{
output A bar plot and a line plot of the comparison between FCS-filtered PROTREC performance
and non-FCS-filtered PROTREC performance.
}
\description{
This performs fcs-filtering of complexes before applying the PROTREC method using the RCC dataset
}
