% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_average_values_n}
\alias{get_average_values_n}
\title{This performs fcs-filtering of complexes before applying the PROTREC method using the RCN dataset}
\usage{
get_average_values_n(FCS_sig_complex_n)
}
\arguments{
\item{FCS_sig_complex_n}{A vector containing proteins associated with significant complexes in FCS}
}
\value{
output A bar plot and a line plot of the comparison between FCS-filtered PROTREC performance
and non-FCS-filtered PROTREC performance.
}
\description{
This performs fcs-filtering of complexes before applying the PROTREC method using the RCN dataset
}
