% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_mat_n}
\alias{get_mat_n}
\title{obtain the FCS significant complexes that overlap with the complexes present in fcs_normal_mat}
\usage{
get_mat_n(replicate)
}
\arguments{
\item{replicate}{the FCS significant complexes derived using a given replicate}
}
\value{
output the those proteins in both FCS results and those that make up complexes in fcs_normal_mat
}
\description{
obtain the FCS significant complexes that overlap with the complexes present in fcs_normal_mat
}
