% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_protprob_normal_mat}
\alias{get_protprob_normal_mat}
\title{obtain the FCS significant complexes that overlap with the complexes present in normal_mat}
\usage{
get_protprob_normal_mat(replicate)
}
\arguments{
\item{replicate}{the FCS significant complexes derived using a given replicate}
}
\value{
output the those proteins in both FCS results and those that make up complexes in normal_mat
}
\description{
obtain the FCS significant complexes that overlap with the complexes present in normal_mat
}
