% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_result_sce_a_hela}
\alias{get_result_sce_a_hela}
\title{This performs hyperparameter tuning of PROTREC based on scenario a using the Hela dataset}
\usage{
get_result_sce_a_hela(hela_ddaprotrec, protrecscoreset, mode)
}
\arguments{
\item{hela_ddaprotrec}{A matrix of PROTREC (1-probabilities) assigned to complexes for the Hela dda dataset}

\item{protrecscoreset}{The threshold of PROTREC score, default 0.95}

\item{mode}{is the mode of PROTREC score selection, default max}
}
\value{
output A matrix with 4 columns and 6 rows: the columns contain information about
the observed proportion of overlap (the PROTREC score), the significance of this overlap (p-value),
the total number of predicted missing proteins, and the number of validated proteins.
The 6 rows represent the 6 permutations of cross-validation using the samples.
}
\description{
This performs hyperparameter tuning of PROTREC based on scenario a using the Hela dataset
}
