% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_result_sce_a_rc}
\alias{get_result_sce_a_rc}
\title{This performs hyperparameter tuning of PROTREC based on scenario a using the RC dataset}
\usage{
get_result_sce_a_rc(rc_nprotrec, rc_cprotrec, protrecscoreset, mode)
}
\arguments{
\item{rc_nprotrec}{A matrix of PROTREC (1-probabilities) assigned to complexes for the RCN dataset}

\item{rc_cprotrec}{A matrix of PROTREC (1-probabilities) assigned to complexes for the RCC dataset}

\item{protrecscoreset}{The threshold of PROTREC score, default 0.95}

\item{mode}{is the mode of PROTREC score selection, default max}
}
\value{
output A matrix with 16 columns and 6 rows: each four columns contain information about
the observed proportion of overlap (the PROTREC score), the significance of this overlap (p-value),
the total number of predicted missing proteins, and the number of validated proteins. The first four columns
correspond to RCN dataset using sample 1 (validated using sample 2), the second four columns correspond to
RCN dataset using sample 2 (validated using sample 1). The next four columns correspond to RCC dataset
using sample 1 (validated using sample 2), the last four columns correspond to RCC dataset using
sample 2 (validated using sample 1). The 6 rows represent the 6 technical replicates of each sample.
}
\description{
This performs hyperparameter tuning of PROTREC based on scenario a using the RC dataset
}
