% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_result_sce_b_hela}
\alias{get_result_sce_b_hela}
\title{This performs hyperparameter tuning of PROTREC based on scenario b using the Hela dataset}
\usage{
get_result_sce_b_hela(
  mode,
  complex_size,
  heladdaprotrec,
  heladdafcs,
  heladdahg
)
}
\arguments{
\item{mode}{is the mode of PROTREC score selection, default max}

\item{complex_size}{is the minimum size of the complexes used}

\item{heladdaprotrec}{A vector of significant proteins for a given sample using PROTREC, default set 0.05 p-value as significant}

\item{heladdafcs}{A vector of significant proteins for a given sample using FCS, default set 0.05 p-value as significant}

\item{heladdahg}{A vector of significant proteins for a given sample using HE, default set 0.05 p-value as significant}
}
\value{
output A matrix with 9 columns and 24 rows: each three columns contain information about
the number N of predicted proteins based on a given method, the results from the given method, and the results from PROTREC.
Each four rows contain information about the observed proportion of overlap (the PROTREC score),
the significance of this overlap (p-value), the total number of predicted missing proteins, and the number of
validated proteins. The six sets of 4 rows that make up the 24 rows correspond to the permutations of
cross-validation using the samples.
}
\description{
This performs hyperparameter tuning of PROTREC based on scenario b using the Hela dataset
}
