% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_result_sce_c_hela}
\alias{get_result_sce_c_hela}
\title{This performs hyperparameter tuning of PROTREC based on scenario c using the Hela dataset}
\usage{
get_result_sce_c_hela(hela_ddaprotrec, protrecscoreset, mode, complex_size)
}
\arguments{
\item{hela_ddaprotrec}{A matrix of PROTREC (1-probabilities) assigned to complexes for the Hela dataset}

\item{protrecscoreset}{The threshold of PROTREC score, default 0.95}

\item{mode}{is the mode of PROTREC score selection, default max}

\item{complex_size}{is the minimum size of the complexes used}
}
\value{
output A matrix with 4 columns and 3 rows: the columns contain information about
the observed proportion of overlap (the PROTREC score), the significance of this overlap (p-value),
the total number of predicted missing proteins, and the number of validated proteins.
The 3 rows represent the validation results using the samples.
}
\description{
This performs hyperparameter tuning of PROTREC based on scenario c using the Hela dataset
}
