% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_result_sce_c_rcn}
\alias{get_result_sce_c_rcn}
\title{This performs hyperparameter tuning of PROTREC based on scenario c using the RCN dataset}
\usage{
get_result_sce_c_rcn(rc_nprotrec, protrecscoreset, mode, size)
}
\arguments{
\item{rc_nprotrec}{A matrix of PROTREC (1-probabilities) assigned to complexes for the RCN dataset}

\item{protrecscoreset}{The threshold of PROTREC score, default 0.95}

\item{mode}{is the mode of PROTREC score selection, default max}

\item{size}{is the minimum size of the complexes used}
}
\value{
output A matrix with 4 columns and 12 rows: the columns contain information about
the observed proportion of overlap (the PROTREC score), the significance of this overlap (p-value),
the total number of predicted missing proteins, and the number of validated proteins.
The 12 rows represent the 12 permutations of cross-validation using the samples.
}
\description{
This performs hyperparameter tuning of PROTREC based on scenario c using the RCN dataset
}
