% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_result_sce_e_hcc}
\alias{get_result_sce_e_hcc}
\title{This performs hyperparameter tuning of PROTREC based on scenario e using the HCC dataset}
\usage{
get_result_sce_e_hcc(hcc_protrec, protrecscoreset, mode)
}
\arguments{
\item{hcc_protrec}{A matrix of PROTREC (1-probabilities) assigned to complexes for the HCC dataset}

\item{protrecscoreset}{The threshold of PROTREC score, default 0.95}

\item{mode}{is the mode of PROTREC score selection, default max}
}
\value{
output A matrix with 4 columns and 15 rows: the columns contain information about
the observed proportion of overlap (the PROTREC score), the significance of this overlap (p-value),
the total number of predicted missing proteins, and the number of validated proteins.
The 15 rows represent the 15 samples.
}
\description{
This performs hyperparameter tuning of PROTREC based on scenario e using the HCC dataset
}
