% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_result_sce_e_lc}
\alias{get_result_sce_e_lc}
\title{This performs hyperparameter tuning of PROTREC based on scenario e using the LC dataset}
\usage{
get_result_sce_e_lc(lc_tprotrec, protrecscoreset, mode)
}
\arguments{
\item{lc_tprotrec}{A matrix of PROTREC (1-probabilities) assigned to complexes for the LC dataset}

\item{protrecscoreset}{The threshold of PROTREC score, default 0.95}

\item{mode}{is the mode of PROTREC score selection, default max}
}
\value{
output A matrix with 8 columns and 19 rows: each four columns contain information about
the observed proportion of overlap (the PROTREC score), the significance of this overlap (p-value),
the total number of predicted missing proteins, and the number of validated proteins.
The first four columns correspond to sample 1 (validated using sample 2), the second four columns
correspond to sample 2 (validated using sample 1). The 19 rows represent the 19 technical replicates.
}
\description{
This performs hyperparameter tuning of PROTREC based on scenario e using the LC dataset
}
