% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_result_sce_e_rcc}
\alias{get_result_sce_e_rcc}
\title{This performs hyperparameter tuning of PROTREC based on scenario e using the RCC dataset}
\usage{
get_result_sce_e_rcc(rc_cprotrec, protrecscoreset, mode)
}
\arguments{
\item{rc_cprotrec}{A matrix of PROTREC (1-probabilities) assigned to complexes for the RCC dataset}

\item{protrecscoreset}{The threshold of PROTREC score, default 0.95}

\item{mode}{is the mode of PROTREC score selection, default max}
}
\value{
output A matrix with 8 columns and 6 rows: each four columns contain information about
the observed proportion of overlap (the PROTREC score), the significance of this overlap (p-value),
the total number of predicted missing proteins, and the number of validated proteins.
The first four columns correspond to sample 1 (validated using sample 2), the second four columns
correspond to sample 2 (validated using sample 1). The 6 rows represent the 6 technical replicates
of each sample.
}
\description{
This performs hyperparameter tuning of PROTREC based on scenario e using the RCC dataset
}
