% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{get_sig_complex}
\alias{get_sig_complex}
\title{Takes the FCS significant complexes and ranks them by the tibtech approach}
\usage{
get_sig_complex(replicate)
}
\arguments{
\item{replicate}{the FCS significant complexes derived using a given replicate}
}
\value{
output the filtered and ranked FCS significant complexes.
}
\description{
Takes the FCS significant complexes and ranks them by the tibtech approach
}
