% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{Hela_ddaProteins}
\alias{Hela_ddaProteins}
\title{This is the Hela_ddaProteins dataset included in my package}
\description{
This is the Hela_ddaProteins dataset included in my package
}
