% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{hela_dia_prot_list2}
\alias{hela_dia_prot_list2}
\title{This is the hela_dia_prot_list2 dataset included in my package}
\description{
This is the hela_dia_prot_list2 dataset included in my package
}
