% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{heladdapepuniq}
\alias{heladdapepuniq}
\title{This is the heladdapepuniq dataset included in my package}
\description{
This is the heladdapepuniq dataset included in my package
}
