% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{hgtest}
\alias{hgtest}
\title{This assigns probabilities to complexes based on HE}
\usage{
hgtest(data, complex_vector, threshold)
}
\arguments{
\item{data}{Proteomics expression data}

\item{complex_vector}{protein complex}

\item{threshold}{the minimum size of the complex required}
}
\value{
output_mat A matrix of HE probabilities assigned to complexes for the dataset
}
\description{
This assigns probabilities to complexes based on HE
}
