% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{hgtest_prot_prob_assign}
\alias{hgtest_prot_prob_assign}
\title{This assigns probabilities to individual proteins based on HE}
\usage{
hgtest_prot_prob_assign(cplx, p)
}
\arguments{
\item{cplx}{complex vector}

\item{p}{is the complex-based HE probability, typically should be the 1 minus the result from certain sample of hgtest}
}
\value{
output The vector of HE probabilities for the proteins of a sample
}
\description{
This assigns probabilities to individual proteins based on HE
}
