% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{mat.ttest}
\alias{mat.ttest}
\title{This performs t-test inside GSEA function}
\usage{
mat.ttest(data)
}
\arguments{
\item{data}{Proteomics expression data}
}
\value{
output_mat A matrix containing all p-values after t-test
}
\description{
This performs t-test inside GSEA function
}
