% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{order_stuff}
\alias{order_stuff}
\title{sort the cumulative distribution}
\usage{
order_stuff(replicate)
}
\arguments{
\item{replicate}{the cumulative distribution of PROTREC scores for the predicted, validated and original proteins}
}
\value{
output the sorted variable
}
\description{
sort the cumulative distribution
}
