% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{pairwise_recovery_protrec}
\alias{pairwise_recovery_protrec}
\title{This works out the recovery significance for PROTREC}
\usage{
pairwise_recovery_protrec(
  prot_predict_list,
  original_prot_list,
  check_prot_list,
  complex_vec,
  protrecscoreset
)
}
\arguments{
\item{prot_predict_list}{A vector of significant proteins for a given sample}

\item{original_prot_list}{The original set of proteins observed for that particular sample}

\item{check_prot_list}{A set of proteins observed in a second replicate}

\item{complex_vec}{A list of complex objects}

\item{protrecscoreset}{The threshold of PROTREC score, default 0.95}
}
\value{
output A vector of five pieces of information: the observed proportion of overlap,
the significance of this overlap (p-value), the number of verified proteins, the total number of predicted missing proteins,
and the list of validated proteins separated by 'a'.
}
\description{
This works out the recovery significance for PROTREC
}
