% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{precision_recall_2}
\alias{precision_recall_2}
\title{This produces the cumulative distribution of PROTREC scores for the predicted, validated and original proteins}
\usage{
precision_recall_2(x, y, z, cplx, g)
}
\arguments{
\item{x}{the probability of proteins observed in the sample}

\item{y}{the names of predicted proteins in the sample}

\item{z}{the names in the check list}

\item{cplx}{the complex information}

\item{g}{the set of proteins in the original list}
}
\value{
output four columns contain information about
the PROTREC score range, the total number of predicted proteins, the number of validated proteins, and
the total number of proteins in the original list.
}
\description{
This produces the cumulative distribution of PROTREC scores for the predicted, validated and original proteins
}
