% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{protprob_cancer_mat}
\alias{protprob_cancer_mat}
\title{This is the protprob_cancer_mat dataset included in my package}
\description{
This is the protprob_cancer_mat dataset included in my package
}
