% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{PROTREC_cplx_prob}
\alias{PROTREC_cplx_prob}
\title{This assigns probabilities to complexes based on PROTREC}
\usage{
PROTREC_cplx_prob(data, complex, fdr, threshold)
}
\arguments{
\item{data}{Proteomics expression data}

\item{complex}{protein complex}

\item{fdr}{false discovery rate of proteomics screen}

\item{threshold}{the minimum size of the complex required}
}
\value{
output_mat A matrix of PROTREC (1-probabilities) assigned to complexes for the dataset
}
\description{
This assigns probabilities to complexes based on PROTREC
}
