% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{RC_ccle}
\alias{RC_ccle}
\title{This is the RC_ccle dataset included in my package}
\description{
This is the RC_ccle dataset included in my package
}
\references{
\url{https://pubmed.ncbi.nlm.nih.gov/31978347/}
}
