% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\name{repgsea}
\alias{repgsea}
\title{This assigns probabilities to complexes based on GSEA}
\usage{
repgsea(data, complex_vector)
}
\arguments{
\item{data}{Proteomics expression data}

\item{complex_vector}{protein complex}
}
\value{
gsea_pvals A matrix of GSEA probabilities assigned to complexes for the dataset
}
\description{
This assigns probabilities to complexes based on GSEA
}
