% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{Siha_ddaProteins}
\alias{Siha_ddaProteins}
\title{This is the Siha_ddaProteins dataset included in my package}
\description{
This is the Siha_ddaProteins dataset included in my package
}
