% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{siha_dia_prot_list3}
\alias{siha_dia_prot_list3}
\title{This is the siha_dia_prot_list3 dataset included in my package}
\description{
This is the siha_dia_prot_list3 dataset included in my package
}
