% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROTREC_Optimization.R
\docType{data}
\name{sihaddapepuniq}
\alias{sihaddapepuniq}
\title{This is the sihaddapepuniq dataset included in my package}
\description{
This is the sihaddapepuniq dataset included in my package
}
